Implement more efficient `with` desugaring
… as permitted by https://github.com/dhall-lang/dhall-lang/pull/1052
