+++ dhall/src/Dhall/Binary.hs
@@ -30,6 +30,7 @@ import Codec.Serialise      (Serialise (decode, encode))
 import Control.Applicative  (empty, (<|>))
 import Control.Exception    (Exception)
 import Data.ByteString.Lazy (ByteString)
+import Data.Foldable (Foldable)
 import Dhall.Syntax
     ( Binding (..)
     , Chunks (..)
@@ -67,6 +68,8 @@ import qualified Data.ByteArray
 import qualified Data.ByteString
 import qualified Data.ByteString.Lazy
 import qualified Data.ByteString.Short
+import qualified Data.Foldable         as Foldable
+import qualified Data.List.NonEmpty    as NonEmpty
 import qualified Data.Sequence
 import qualified Dhall.Crypto
 import qualified Dhall.Map
@@ -580,6 +583,23 @@ decodeExpressionInternal decodeEmbed = go
 
                                 return (ListLit (Just _T) empty)
 
+                            29 -> do
+                                l <- go
+
+                                n <- Decoding.decodeListLen
+
+                                ks₀ <- replicateDecoder n Decoding.decodeString
+
+                                ks₁ <- case NonEmpty.nonEmpty ks₀ of
+                                    Nothing ->
+                                        die "0 field labels in decoded with expression"
+                                    Just ks₁ ->
+                                        return ks₁
+
+                                r <- go
+
+                                return (With l ks₁ r)
+
                             _ ->
                                 die ("Unexpected tag: " <> show tag)
 
@@ -940,8 +960,12 @@ encodeExpressionInternal encodeEmbed = go
                 (go t)
                 (go _T)
 
-        a@With{} ->
-            go (Syntax.desugarWith a)
+        With l ks r ->
+            encodeList4
+                (Encoding.encodeInt 29)
+                (go l)
+                (encodeList (fmap Encoding.encodeString ks))
+                (go r)
 
         Note _ b ->
             go b
@@ -975,11 +999,12 @@ encodeList4 :: Encoding -> Encoding -> Encoding -> Encoding -> Encoding
 encodeList4 a b c d = Encoding.encodeListLen 4 <> a <> b <> c <> d
 {-# INLINE encodeList4 #-}
 
-encodeListN :: Int -> [ Encoding ] -> Encoding
-encodeListN len xs = Encoding.encodeListLen (fromIntegral len) <> mconcat xs
+encodeListN :: Foldable f => Int -> f Encoding -> Encoding
+encodeListN len xs =
+    Encoding.encodeListLen (fromIntegral len) <> Foldable.fold xs
 {-# INLINE encodeListN #-}
 
-encodeList :: [ Encoding ] -> Encoding
+encodeList :: Foldable f => f Encoding -> Encoding
 encodeList xs = encodeListN (length xs) xs
 {-# INLINE encodeList #-}
